NOJIRA: Create common SDL class
This allows us to eventually create other programs that can help make
the game like a Map Editor or something else.

A map editor would be useful to visually see the map that will be used
in the game. We can convert that map to something that can be parsed
and understood by the game.