Game+MapEditor: Lots of changes
In theory the MapEditor can save a map and it'll render in the
game. However, the MapEditor doesn't load the map or allow the user to
delete sprites either.

The Game scaled it's sprites 3.125x but lost it's correct camera
implementation, there needs to be a scalable refactor to this.

Sprite overlap is in action to avoid having blank drawing underneath
transparent sprites making the game look more natural.

Player animation was somewhat fixed.

Fixed memory leak issue with FPS texture.