Game: Experiment with Back Buffer side-scroll
The idea would be to pre-allocate a back buffer to the map itself.
Therefore there would need to be some kind of pre-fetching or
pre-allocation of buffers as the player gets nearer a map they haven't
seen before.

The back buffer itself trivially solves side-scrolling by manipulating
the source rect on the buffer.

This change intentionally regresses other things such as attack
animation and so on since this is a work in progress. Those
regressed features will be added back later.

The mechanism by which we perform the render loop changes now too to
wait for events depending on last draw. This allows us to maintain a
healthy 60fps without spin-locking.