Game: Use new back-buffer mechanism
We keep the back buffer size at the viewport. Then we blit depending
on player direction to "fill" the new contents of the map. After blit
we update the back buffer and then use that to render.

This greatly simplifies rendering the map in general.

We also upscale the spritesheet to begin with to make blitting easier.