MapEditor: Iterate on "MARK" mode.
For now, mark mode will draw a border around top-left of a 4x4 map
tile and then store this state in a dynamic array. The next idea is on
save to binary search through placements until the marked region is
fully searched and then to save this into some Map.h. However, this
remains unfinished for now. There's a missing implementation on binary
searching entire rows which would allow this to be done in a quick manner.

There's also the slight issue of overlapping marked regions that is
not solved but also not a blocker for the MapEditor.

We also pulled in tsoding's Arena implementation since we'll likely
move to SDL2 and C in the future and it makes memory management far
easier down the line.