NOJIRA: BackBuffer for MapEditor
Now it's possible to move the Placement Grid by doing
UP/DOWN/LEFT/RIGHT, however the heavy lifting required to draw the
next strips is not there and it regresses drag&drop. There are FIXMEs
to plan to adjust that but looks good so far.

In theory this means the MapEditor should perform better given that we
aren't doing redundant draws anymore. Moving the placement grid is
necessary to construct non-trivially complex maps going forward.