NOJIRA: Scroll the map editor
We can now draw horizontal/vertical snap lines and move the placement
area around.

However, now when we try to delete sprites, we need to account for
camera_x and camera_y.

We also need to take into account any sprites that might be in that
strip. Which means when loading in the map, we might have to perhaps
load co-ordinates to somewhere and respect them when scrolling around.

We should also probably demarcate what the map boundaries are on load.

Finally, the encoding we have today for the map needs to extend to
support multiple maps in general.