NOJIRA: Bunch of changes - too many to enumerate
Try to move to searching for Entities by tags, this simplifies logic a
lot more. For example when we want to draw the next MapEditor strip in
the Placement Area, we want to look for entities in that row. Then
subsequently, want to find the right texture to draw by filtering for
that entity's tag.

We partially managed to write some of the implementation for drawing
the next strip and keeping entities around visually speaking.